Fix CVE-2017-17426
The malloc function in the GNU C Library (aka glibc or libc6) 2.26 could
return a memory block that is too small if an attempt is made to
allocate an object whose size is close to SIZE_MAX, potentially leading
to a subsequent heap overflow. This occurs because the per-thread cache
(aka tcache) feature enables a code path that lacks an integer overflow
check.

Signed-off-by: Victor Rodriguez <victor.rodriguez.bahena@intel.com>