Address performance issues with LD_AUDIT (rtld-audit)
This commit includes a patch which will help the performance of
linux-steam-integration through use of the rtld-audit interface,
via LD_AUDIT.

Specifically it omits profiling for LD_AUDIT modules which do not
expose la_plt hooks, thus not impacting performance. Without this
patch, we would expect performance to be degraded when using LSI,
the opposite of our intention.

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>