Add -mstackrealign to the 32-bit build
This adds a performance penalty to any and all entry points in the
32-bit glibc, but is necessary because there's a lot of software (games,
especially) we're trying to be compatible with that hasn't been rebuilt
in 15 years. That software is still using the old i386 ABI, which only
required 4-byte aligned stacks, whereas anything from the past 15 years
uses 16-byte aligned stacks.

The performance penalty is acceptable because this is 32-bit (it's not
performance critical in the first place).

Fixes clearlinux/distribution#1152 (unverified)